backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
fastapi==0.120.4
idna==3.11
mangum==0.19.0
pydantic==2.12.3
pydantic_core==2.41.4
sniffio==1.3.1
starlette==0.49.1
typing-inspection==0.4.2
typing_extensions==4.15.0
requests==2.32.5
google-auth==2.42.1
uvicorn==0.38.0
python-dotenv==1.0.0